Reviewed-on: http://git.bitilinks.cn/bitilinks/ydme-ems/pulls/618

+++ ems-module-power/ems-module-power-biz/src/main/java/cn/bitlinks/ems/module/power/controller/admin/report/supplywatertmp/vo/LtwtChartVO.java
@@ -22,10 +22,10 @@ public class LtwtChartVO {
     private List<BigDecimal> ydata2;
 
     @Schema(description = "上限")
-    private Integer max;
+    private BigDecimal max;
 
     @Schema(description = "下限")
-    private Integer min;
+    private BigDecimal min;
 
     @Schema(description = "x轴")
     private List<String> xdata;
